fix : 회원 상벌점, 기수 추가로 인해 생긴 테스트 오류 수정
issue : #157 #158 #162
