+++ src/test/java/keeper/project/homepage/controller/etc/AboutContentControllerTest.java
@@ -93,6 +93,7 @@ public class AboutContentControllerTest extends ApiControllerTestSetUp {
         .memberType(memberTypeEntity)
         .memberRank(memberRankEntity)
         .thumbnail(thumbnailEntity)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberAdminJobEntity)))
         .build();
     memberRepository.save(memberAdmin);
