+++ src/test/java/keeper/project/homepage/controller/etc/AboutTitleControllerTest.java
@@ -108,6 +108,7 @@ public class AboutTitleControllerTest extends ApiControllerTestSetUp {
         .memberType(memberTypeEntity)
         .memberRank(memberRankEntity)
         .thumbnail(thumbnailEntity)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberAdminJobEntity)))
         .build();
     memberRepository.save(memberAdmin);
