+++ src/test/java/keeper/project/homepage/controller/library/BookManageControllerTest.java
@@ -160,6 +160,7 @@ public class BookManageControllerTest extends ApiControllerTestSetUp {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
@@ -175,6 +176,7 @@ public class BookManageControllerTest extends ApiControllerTestSetUp {
         .nickName(adminNickName)
         .emailAddress(adminEmailAddress)
         .studentId(adminStudentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberAdminJobEntity)))
         .build();
     memberRepository.save(memberAdmin);
