+++ src/test/java/keeper/project/homepage/controller/library/LibraryMainControllerTest.java
@@ -93,6 +93,7 @@ public class LibraryMainControllerTest extends ApiControllerTestSetUp {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
