+++ src/test/java/keeper/project/homepage/controller/sign/SignUpControllerTest.java
@@ -62,6 +62,7 @@ public class SignUpControllerTest extends ApiControllerTestSetUp {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
