+++ src/test/java/keeper/project/homepage/repository/member/MemberHasCommentDislikeRepositoryTest.java
@@ -91,6 +91,7 @@ public class MemberHasCommentDislikeRepositoryTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
