+++ src/test/java/keeper/project/homepage/repository/member/MemberRepositoryTest.java
@@ -62,6 +62,7 @@ public class MemberRepositoryTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
@@ -100,6 +101,7 @@ public class MemberRepositoryTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(member);
