+++ src/test/java/keeper/project/homepage/repository/posting/CommentRepositoryTest.java
@@ -78,6 +78,7 @@ public class CommentRepositoryTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
