+++ src/test/java/keeper/project/homepage/service/ThumbnailServiceTest.java
@@ -130,6 +130,7 @@ public class ThumbnailServiceTest extends ApiControllerTestSetUp {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
