+++ src/test/java/keeper/project/homepage/service/member/MemberHasCommentDislikeServiceTest.java
@@ -88,6 +88,7 @@ public class MemberHasCommentDislikeServiceTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
