+++ src/test/java/keeper/project/homepage/service/member/MemberHasCommentLikeServiceTest.java
@@ -89,6 +89,7 @@ public class MemberHasCommentLikeServiceTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
