+++ src/test/java/keeper/project/homepage/service/member/MemberServiceTest.java
@@ -54,6 +54,7 @@ public class MemberServiceTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
