+++ src/test/java/keeper/project/homepage/service/posting/CommentServiceTest.java
@@ -98,6 +98,7 @@ public class CommentServiceTest {
         .nickName(nickName)
         .emailAddress(emailAddress)
         .studentId(studentId)
+        .generation(0F)
         .memberJobs(new ArrayList<>(List.of(hasMemberJobEntity)))
         .build();
     memberRepository.save(memberEntity);
